+++ src/haven/preprocessors/__init__.py
@@ -1,0 +1,4 @@
+from .shutter_suspender import shutter_suspend_decorator, shutter_suspend_wrapper
+from .baseline import baseline_wrapper, baseline_decorator
+from .inject_metadata import inject_haven_md_wrapper
+from .auto_shutter import auto_shutter_wrapper, auto_shutter_decorator
